Dodata nova polja u model

+++ Backend/Models/Comment.cs
@@ -5,15 +5,15 @@ namespace Backend.Models
 
     public class Comment
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Content { get; set; }
-        public int IssueId { get; set; }
+        public long IssueId { get; set; }
         public Issue Issue { get; set; }
-        public int PullRequestId { get; set; }
+        public long PullRequestId { get; set; }
         public PullRequest PullRequest { get; set; }
-        public int DeveloperId { get; set; }
+        public long DeveloperId { get; set; }
         public Developer Developer { get; set; }
-        public int ReactionId { get; set; }
+        public long ReactionId { get; set; }
         public Reaction Reaction { get; set; }
     }
